Week3/Day4/DailyChallenge/requirements.txt
certifi==2024.7.4
charset-normalizer==3.3.2
distlib==0.3.8
filelock==3.15.4
idna==3.7
platformdirs==4.2.2
psycopg2==2.9.9
python-dotenv==1.0.1
requests==2.32.3
urllib3==2.2.2
virtualenv==20.26.3
